pyproject.toml
[tool.black]
line-length = 120

[tool.mypy]
disallow_untyped_defs = true
show_error_codes = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
explicit_package_bases = true
exclude = [".MOCBOT"]

[tool.poetry]
name = "mocbot-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-ext-menus = "1.1.0"
discord-py = "2.3.2"
expiringdict = "1.2.2"
lavalink = "5.7.1"
pillow = "9.3.0"
python-socketio = "5.11.0"
pyyaml = "6.0.1"
requests = "2.31.0"
shortuuid = "1.0.13"
spotifysearch = "0.0.5"
stringprogressbar = "1.1.1"
ytmusicapi = "1.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
